+++ src/components/application-loader/initializers/configLoader.ts
@@ -2,7 +2,7 @@ import { getConfig } from '../../../api/config'
 import { setApiUrl } from '../../../redux/api-url/methods'
 import { setBanner } from '../../../redux/banner/methods'
 import { setConfig } from '../../../redux/config/methods'
-import { getAndSetUser } from '../../../api/utils'
+import { getAndSetUser } from '../../login-page/auth/utils'
 
 export const loadAllConfig: (baseUrl: string) => Promise<void> = async (baseUrl) => {
   setApiUrl({
