+++ src/components/login-page/auth/via-ldap.tsx
@@ -5,7 +5,7 @@ import { Trans, useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { doLdapLogin } from '../../../api/auth'
 import { ApplicationState } from '../../../redux'
-import { getAndSetUser } from '../../../api/utils'
+import { getAndSetUser } from './utils'
 
 export const ViaLdap: React.FC = () => {
   const { t } = useTranslation()
