+++ src/components/login-page/auth/via-openid.tsx
@@ -2,7 +2,7 @@ import React, { FormEvent, useState } from 'react'
 import { Alert, Button, Card, Form } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
 import { doOpenIdLogin } from '../../../api/auth'
-import { getAndSetUser } from '../../../api/utils'
+import { getAndSetUser } from './utils'
 
 export const ViaOpenId: React.FC = () => {
   useTranslation()
